+++ CHANGELOG.md
@@ -6,7 +6,19 @@ Gotta blame the broken mods!
 ------------------------------------------------
        | Blame Fabric changelog |
        
-          
+      
+   (V.2.2.7 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)
+      
+    Mob Spawning:
+•  Will detect if mob spawning will crash due to mob entries having a weight of 0 or negative value.
+
+    Dispenser Behaviors:
+•  Fixed potential crash due to Dispenser registration not being syncronized by Mojang which makes it harder for Blame to work properly.
+
+    Missing Loottable:
+•  Fixed jeresources check not working to help reduce some logspam.
+
+
    (V.2.2.6 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)
       
     Dispenser Behaviors: